Tag 2023-4: First working image

+++ src/ILogParser.go
@@ -19,7 +19,7 @@ type HttpRequest struct {
 	mobile         bool
 	country        string
 	city           string
-	latency        float64
+	latency        int
 	xForwardedFor  string
 }
